Add files via upload
Hi guys, Amanda filled me in on all the stuff you've been doing (awesome!), and I was able to fully implement the PC <- R_b and R_b <- PC+1 portion of call and return. As of right now though, the Frame Pointer is being ignored. Hence, the microprocessor can run HERA programs that DO NOT have nested function calls (i.e. if we pretend the stack is one frame, then it works!!). So, every test but the 19th one passes from the 246 test suite (but this is expected, as the 19th test is a recursive power function that most certainly has nested calls). Let me know via email or group chat if you guys have any questions!