keyboard 2 working but with bug
if you type "abc"; keyboard stores a and b correctly in memory. however once it finishes storing b and jumps back to the infinite loop (it's at 0x000a in ROM at this point which holds "100b") the BR in branching control says it should jump to somewhere else instead of the top of the loop (0x0004 in ROM). The READ2 value is wrong and I'm unsure why.